bot_for_partner/serializers/serializers_owner.py
from rest_framework import serializers

from bot_for_partner.models import Owner


class OwnerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Owner
        fields = '__all__'
