requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
pillow==10.4.0
python-telegram-bot==21.5
sniffio==1.3.1
sqlparse==0.5.1
